Restrict tokens to applications instead of projects
